+++ include/managers/views/wardriving_screen.h
@@ -1,0 +1,12 @@
+#ifndef WARDRIVING_SCREEN_H
+#define WARDRIVING_SCREEN_H
+
+#include "managers/display_manager.h"
+
+extern View wardriving_view;
+
+void wardriving_view_create(void);
+void wardriving_view_destroy(void);
+void wardriving_view_set_scan_mode(bool enabled);
+
+#endif